Refactored Virtual machine to remove data structure specific functions to their own packages
Linting
